[Feat] Configuring a network environment that can be run for 4 people
- 4인 기준 실행 네트워크 환경 구성
